+++ e2e/helpers/playwright/index.ts
@@ -1,2 +1,1 @@
-export * from './fixture';
 export * from './with-isolated-page';
