+++ e2e/tests/admin/analytics/overview.test.ts
@@ -1,4 +1,6 @@
-import {test, expect, withIsolatedPage} from '../../../helpers/playwright';
+import {test, expect} from '../../../helpers/playwright/fixtures/authenticated-fixture';
+import {withIsolatedPage} from '../../../helpers/playwright';
+
 import {
     AnalyticsOverviewPage,
     AnalyticsWebTrafficPage,
