+++ e2e/tests/admin/tags/editor.test.ts
@@ -1,4 +1,4 @@
-import {test, expect} from '../../../helpers/playwright';
+import {test, expect} from '../../../helpers/playwright/fixtures/authenticated-fixture';
 import {TagsPage, TagEditorPage, NewTagsPage} from '../../../helpers/pages/admin';
 
 test.describe('Ghost Admin - Tags Editor', () => {
